refactor(frontend): collapse the three byte-identical zoom platform forks into one zoom.ts
Phase 3 of docs/reports/REPORT_DRY_AND_MODULARITY_AUDIT_2026_09_06.md
(§2.3). zoom.{win32,linux,darwin}.ts were selected per-platform by the
`.platform` import resolver, but measured on 2026-09-06 they differed by
nine lines — every one a comment. The JavaScript was identical on all
three platforms.

The per-platform knowledge those comments carried is real, so the merged
file's header keeps all three rationales (Windows: calc(100vw/zoom) in
CSS; macOS: width:100% and do NOT use the calc form on WebKit; Linux:
WebKitGTK needs no compensation) and the one rule they share — JS sets
only --zoomfactor; width compensation stays in CSS.

- frontend/app/store/zoom.ts: the win32 body verbatim + merged header.
- Deleted zoom.win32.ts / zoom.linux.ts / zoom.darwin.ts, and the
  zoom.platform.ts tsc typecheck stub that re-exported ./zoom.win32
  (now dead: every import targets zoom directly).
- Rewrote the six `zoom.platform` import sites (five static, one dynamic
  in app-init.ts) to `zoom`, and two comment references.

The `.platform` resolver is untouched and still serves the four families
that genuinely differ (TileLayout, CrossWindowDragMonitor, useWindowDrag,
window-controls). Zero behavior change.

Verified: tsc --noEmit clean; vitest over frontend/app/store,
frontend/app/element, frontend/app/view/helpview -> 823 passed (29 files).
