+++ .changesets/1788731115-refactor-frontend-collapse-the-three-byte-identical-zoom-win-d6r5.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(frontend): collapse the three byte-identical zoom.{win32,linux,darwin}.ts platform forks into one zoom.ts
