--- frontend/app/store/zoom.platform.ts
@@ -1,7 +1,0 @@
-// Copyright 2026, AgentMux Corp.
-// SPDX-License-Identifier: Apache-2.0
-//
-// TypeScript type-checking stub — resolved to zoom.{win32,darwin,linux}.ts
-// at build time by Vite's platformResolve plugin.
-// Do NOT import this file directly in application code.
-export * from "./zoom.win32";
